requirements.txt

# Reporting and logging
allure-pytest>=2.13.2

# Development dependencies
black>=23.9.1
faker>=19.6.2
flake8>=6.1.0

# Optional: For API testing integration
httpx>=0.25.0
isort>=5.12.0
mypy>=1.6.0
pillow>=10.0.1
pre-commit>=3.4.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
# Core testing dependencies
pytest>=7.4.0
pytest-html>=3.2.0
pytest-rerunfailures>=12.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1

# Utilities and helpers
python-dotenv>=1.0.0
requests>=2.31.0
selene>=2.0.0rc4

# Selenium and browser automation
selenium>=4.15.0
structlog>=23.1.0
webdriver-manager>=4.0.1
